패키지 구조 통일 작업
