+++ src/main/java/com/numble/whatz/application/video/controller/VideoController.java
@@ -1,11 +1,10 @@
-package com.numble.whatz.web.video;
+package com.numble.whatz.application.video.controller;
 
-import com.numble.whatz.web.home.VideoInfoDto;
+import com.numble.whatz.application.video.controller.dto.*;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
