+++ src/main/java/com/numble/whatz/application/video/controller/dto/DirectDto.java
@@ -1,4 +1,4 @@
-package com.numble.whatz.web.video;
+package com.numble.whatz.application.video.controller.dto;
 
 import lombok.Getter;
 import lombok.Setter;
