+++ src/main/java/com/numble/whatz/application/video/controller/dto/VideoOneDto.java
@@ -1,29 +1,28 @@
-package com.numble.whatz.web.home;
+package com.numble.whatz.application.video.controller.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Getter @Setter
-public class VideoInfoDto {
+public class VideoOneDto {
 
-    private String nickname;
-    private String profile;
     private int likes;
     private String title;
     private LocalDateTime videoDate;
     private long views;
     private String videoThumbnail;
+    private String linkOrPath;
+    private String content;
 
-    public VideoInfoDto(String nickname, String profile, int likes, String title, LocalDateTime videoDate, long views, String videoThumbnail) {
-        this.nickname = nickname;
-        this.profile = profile;
+    public VideoOneDto(int likes, String title, LocalDateTime videoDate, long views, String videoThumbnail, String linkOrPath, String content) {
         this.likes = likes;
         this.title = title;
         this.videoDate = videoDate;
         this.views = views;
         this.videoThumbnail = videoThumbnail;
+        this.linkOrPath = linkOrPath;
+        this.content = content;
     }
 }
